requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg==3.2.12
psycopg-binary==3.2.12
python-dotenv==1.0.0

